Move all `image2tensor` code to `lib.rs`
With its small size, there is no need for a separate module.
